feat: `FileManagmentService` python examples
